refactor: extract helper functions
#2294
